♻️ refactor: useEducation 훅 분리 #429
- form을 담당하는 useEducationForm.ts와 api를 담당하는 useEducationApi.ts로 분리
- 전공 enum관련 타입 매칭 이슈도 같이 해결
